Move figure code to figure.js
